ai/requirements.txt
fastapi>=0.111.0,<0.116.0
uvicorn[standard]>=0.30.0,<0.32.0
python-dotenv>=1.0.1,<2.0.0
langchain>=0.2.12,<0.3.0
langchain-core>=0.2.26,<0.3.0
langchain-community>=0.2.12,<0.3.0
langchain-google-genai>=1.0.4,<2.0.0
langchain-text-splitters>=0.0.1,<1.0.0
pymilvus>=2.4.4,<3.0.0
sqlalchemy>=2.0.30,<3.0.0
mysql-connector-python>=8.3.0,<9.0.0
pydantic>=2.8.0,<3.0.0
pytest>=8.3.0,<9.0.0
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0
pymysql>=1.1.0,<2.0.0
mlxtend>=0.23.0,<1.0.0
